requirements.txt
# Flood Estimation Project - Requirements
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Segmentation Models
segmentation-models-pytorch>=0.3.3

# Data Augmentation
albumentations>=1.3.0
opencv-python>=4.8.0

# GIS & Geospatial Processing
rasterio>=1.3.0
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
fiona>=1.9.0

# Terrain Analysis (for flow direction and DEM processing)
whitebox>=2.3.0
richdem

# Core Scientific Libraries
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Utilities
tqdm
pyyaml
pillow

# Jupyter (for notebooks)
jupyter
ipywidgets
jupytext  # For version control of notebooks as .py files
